promoted landing page to index.tsx
